WIP: Add tests for the StatefulSet controllers
The statefulSetServingCertUpdateController tests are directly modelled
after the existing tests for (only) the refresh logic of
serviceServingCertUpdateController; the StatefulSet controller only
uses the beta annotations, so we have one test instead of two.

The statefulSetServingCertController tests are similarly directly
modelled after tests for serviceServingCertController , and only use
the beta annotations.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>