Fix failures on repeated e2e tests in the same cluster
The spec.unsupportedConfigOverrides forceRotation field
only reacts to changes of value, so instead of always writing 42
there, write a value different from the last one.

The "infinite" loop will have at most 2 iterations because
we are comparing the generated name with a single value of
an annotation from a loaded secret. (That's racy against concurrent
runs of the same e2e test suite, but the test suite is already assuming
that it's the only code changing the openshift-service-ca{,-operator}'s
state.)

Signed-off-by: Miloslav Trmač <mitr@redhat.com>