Don't abort metrics/service-ca-metrics before the sync finishes
Don't immediately abort when we observe a 0 value of
service_ca_expiry_time_seconds, which is only updated inside
the operator's sync loop.

Keep trying for the full timeout value.

It's tempting to argue that the metric should not be reported as 0
until the value is known, OTOH
https://prometheus.io/docs/practices/instrumentation/#avoid-missing-metrics
says "export a default value such as 0 for any time series you know may
exist in advance."

Signed-off-by: Miloslav Trmač <mitr@redhat.com>