Dump data of failing checks in test/util/rotate.go
If these tests occasionally fail, it's impossible to diagnose
that failure without capturing the full data (e.g. to tell
whether the data is just empty or corrupt, or whether there
really is an unexpected intermediate state where the
verification fails). So include the full failing data in the
error message.

Originally motivated by a
>     rotate.go:151: Failed to receive output: Get "https://test-service-vyh1o.test-urr50.svc:62133": x509: certificate signed by unknown authority (possibly because of "crypto/rsa: verification error" while trying to verify candidate authority certificate "openshift-service-serving-signer@1615214293")

which I have only seen once; hopefully this will help if it ever reoccurs.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>